Refactor Settings shared helpers
